Revert "Use explicit type declarations for uint256"
This reverts commit fcff1cfcbe118581a01f09b0abf659e4d7b4adae.

+++ contracts/EIP777/IEIP777TokensSender.sol
@@ -5,7 +5,7 @@ interface IEIP777TokensSender {
     address operator,
     address from,
     address to,
-    uint256 amount,
+    uint amount,
     bytes userData,
     bytes operatorData
   ) public;
