+++ contracts/wallet/TestCalls.sol
@@ -5,13 +5,13 @@ pragma solidity ^0.4.15;
 contract TestCalls {
 
   // msg.data.length of the latest call to "receive" methods
-  uint256 public lastMsgDataLength;
+  uint public lastMsgDataLength;
 
   // msg.value of the latest call to "receive" methods
-  uint256 public lastMsgValue;
+  uint public lastMsgValue;
 
-  uint256 public uint1;
-  uint256 public uint2;
+  uint public uint1;
+  uint public uint2;
   bytes public byteArray1;
 
   modifier setMsgFields {
@@ -24,11 +24,11 @@ contract TestCalls {
     // This constructor will be used to test the creation via multisig wallet
   }
 
-  function receive1uint(uint256 a) setMsgFields payable public {
+  function receive1uint(uint a) setMsgFields payable public {
     uint1 = a;
   }
 
-  function receive2uints(uint256 a, uint256 b) setMsgFields payable public {
+  function receive2uints(uint a, uint b) setMsgFields payable public {
     uint1 = a;
     uint2 = b;
   }
